requirements-linux.txt
# Linux-compatible requirements for AgentCore container deployment
# Excludes Windows-specific packages like pywin32

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
authlib==1.6.5
aws-requests-auth==0.4.3
awscrt==0.29.2
beartype==0.22.8
beautifulsoup4==4.14.3
bedrock-agentcore==1.1.1
boto3==1.42.5
botocore==1.42.5
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.3.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.1
fastmcp==2.13.2
filelock==3.20.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
identify==2.6.15
idna==3.11
importlib-metadata==8.7.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
log-symbols==0.0.14
markdown-it-py==4.0.0
markdownify==1.2.2
mcp==1.23.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
nodeenv==1.9.1
ollama==0.6.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-threading==0.60b0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
pillow==11.3.0
platformdirs==4.5.1
pre-commit==4.5.0
prompt-toolkit==3.0.52
propcache==0.4.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
# pywin32==311  # Windows-only package - excluded for Linux container
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.16.0
shellingham==1.5.4
six==1.17.0
slack-bolt==1.27.0
slack-sdk==3.39.0
soupsieve==2.8
spinners==0.0.24
sqlalchemy==2.0.45
sse-starlette==3.0.3
starlette==0.50.0
strands-agents==1.19.0
strands-agents-builder==0.1.10
strands-agents-tools==0.2.17
sympy==1.14.0
tenacity==9.1.2
teradata-mcp-server==0.1.6
teradatasql==20.0.0.47
teradatasqlalchemy==20.0.0.8
termcolor==3.2.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
virtualenv==20.35.4
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0